Implement pagination
There is a list of tickers with its own properties which serves as an
index for paginating or cycling through different stocks.

Every X seconds the page is switched and so the new ticker selected

IMPORTANT: while a ticker is loading, there might be errors and the old
ticker stays on display. The HTTP request is blocking, therefore the
time in between pages can (and most-likely will) exceed the the
REFRESH_INTERNVAL variable...